--- backend/src/main/java/minionz/backend/issue/model/IssueStatus.java
@@ -1,8 +1,0 @@
-package minionz.backend.issue.model;
-
-public enum IssueStatus {
-    NO_STATUS,
-    TODO,
-    IN_PROGRESS,
-    DONE
-}
