+++ backend/src/main/java/minionz/backend/scrum/issue/model/IssueStatus.java
@@ -1,0 +1,8 @@
+package minionz.backend.scrum.issue.model;
+
+public enum IssueStatus {
+    NO_STATUS,
+    TODO,
+    IN_PROGRESS,
+    DONE
+}
