+++ backend/src/main/java/minionz/backend/scrum/sprint/model/SprintStatus.java
@@ -1,0 +1,7 @@
+package minionz.backend.scrum.sprint.model;
+
+public enum SprintStatus {
+    TODO,
+    IN_PROGRESS,
+    DONE
+}
