+++ backend/src/main/java/minionz/backend/scrum/sprint_label_select/model/SprintLabelSelect.java
@@ -1,0 +1,28 @@
+package minionz.backend.scrum.sprint_label_select.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+import minionz.backend.scrum.sprint.model.Sprint;
+import minionz.backend.scrum.sprint_label.model.SprintLabel;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Entity
+public class SprintLabelSelect {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long sprintLabelSelectId;
+
+    // SprintLabelSelect : SprintLabel = N : 1
+    @ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @JoinColumn(name = "sprint_label_id")
+    private SprintLabel sprintLabel;
+
+    // SprintLabelSelect : Sprint = N : 1
+    @ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @JoinColumn(name = "sprint_id")
+    private Sprint sprint;
+}
