--- backend/src/main/java/minionz/backend/sprint/model/SprintStatus.java
@@ -1,7 +1,0 @@
-package minionz.backend.sprint.model;
-
-public enum SprintStatus {
-    TODO,
-    IN_PROGRESS,
-    DONE
-}
